Fixes from Daniel Pittman <daniel@danann.net>:
* Fixed 'rcp-run-real-handler'

This removes the lossage of my previous patch, replacing it with a much
more useful version. This one disables EFS and RCP for the selected
operation and looks almost identical to the one that the documentation
recommends.


* Remove 'vc-registered' from 'rcp-file-name-handler-alist'

After fixing the lossage I suggested for 'rcp-run-real-handler' I went
to work fixing remote VC for XEmacs. As part of this,
'rcp-handle-vc-registered' became a no-op, calling directly to the
default handler.

Since this is the behavior for an unknown function, I removed
'vc-registered' from the alist.


* Fix the window behavior of 'rcp-vc-do-command'

This did not preserve the window layout, resulting in odd changes to the
screen layout. This now preserves that layout where the changes were
happening. As a result, remote VC feels much more like local VC.


* Implement 'rcp-vc-simple-command'

XEmacs uses both 'vc-simple-command' and 'vc-do-command' for VC, but
only 'vc-do-command' was modified by RCP. I implemented
'rcp-vc-simple-command' based on the 'rcp-vc-do-command' template and
the XEmacs VC code.


As a result of this and the 'rcp-run-real-handler' changes, VC now works
remotely where the local and remote login names are identical. I will
take a look at the 'vc-user-login-name' function soon and see if I can
fix the remote/local mismatched name issues.


* Added (sit-for 1) to 'rcp-post-connection'

This is needed, for me at least, when using the 'scp' method between my
Linux box and several different remote Linux machines.


* Expanded on the comment before my advice on 'efs-ftp-path'

You had put a comment in there with a question - now that question is
answered.