Initial infrastructure for implementing remote file information caching.
This is *not* a working implementation or anything like that. It's a start at
providing the background support needed to put caching into the main body of
the code.

I intend to flesh out this infrastructure and begin to use it through the
code over the next few days.

This is being made available to solicit comments on the public interface and
the nature of the data storage (if you can work it out :)