Performance work:
Use grep(1) to filter the filenames on the remote machine when doing filename
completion. This should (substantially) reduce the network traffic when you
have a partial filename and perform completion in a large directory.

If grep(1) cannot be found on the remote machine(!), it will fall back to the
old behavior of doing local filtering.