`insert-file-contents' and `write-region' now work.
TRAMP2 is now hooked into the `file-name-handler-alist' and basic tramp2 file
operations work successfully. File content is actually visited.

The system is basically functional but not ready for production. It requires
substantially more functional completeness before it can be used in a
real-world situation.

There is a current known bug with version control integration when visiting
files. This is being looked at. If you want to test this, expect that.