Merge from trunk:
* configure.ac: Bump version (to 2.0.41) and set the
last-merged-with-Emacs version (to 2.0.40).  Hm.  The version bump
to 2.0.41 is probably bogus, we want to create a stable
branch...  Argh.

* Makefile.in (emacs): Don't patch man/trampver.texi in the Emacs
repository, just copy texi/trampver.texi there.  We can't run
"cvs diff" on texi/trampver.texi since it is a generated file, so
our patching trick won't work for this file.

* configure.ac (TRAMP_EMACS_MERGE_VERSION): New variable,
indicating which was the last revision of Tramp that was merged
into Emacs.  Don't forget to update this after running "make
emacs" (or "gmake emacs")!

* Makefile.in (EMACS_EL_FILES): Mention files explicitly to make
sure that tramp-efs.el is not included.
(EMACS_TEXI_FILES): New variable, lists manual files that need to
be merged with the Emacs repository.
(TRAMP_EMACS_MERGE_VERSION, TRAMP_EMACS_MERGE_TAG): New
variables, indicating which was the last revision of Tramp that
was merged into Emacs.  Automatically set by configure, original
is in configure.ac.
(emacs): Reword to use patches instead of copying the files.
This helps to make sure that changes to Tramp files in the Emacs
repository aren't overwritten.