* Makefile.in (savannah): Don't apply "configure" for Japanese manual, but call "make -C texi" with proper variable settings.
* aclocal.m4 (AC_EMACS_INFO): Write default values in help output
for "--with-xemacs" and "--with-emacs".
(AC_PATH_LISPDIR): Write default value in help output for
"--with-lispdir".
(all): Unify capitalization for help output.

* configure.ac: Bump version (to 2.1.3-pre).

* lisp/tramp.el (top): Apply `def-edebug-spec' only if function is
defined.  This is not the case for XEmacs without package
"edebug".
(tramp-set-auto-save-file-modes): Set permissions of autosaved
remote files to the permissions of the original file.  This is not
the case for Emacs < 21.3.50 and XEmacs < 21.5.  Add function to
`auto-save-hook'.  Reported by Thomas Prokosch <thomas@nadev.net>.

* texi/Makefile.in (NONGNU): Kai uses a new directory structure in his
working copies.
(savannah): Copy also index files.

* texi/tramp.texi (Frequently Asked Questions): Comment paragraph about
plink link.  The URL is outdated.  Originator contacted for
clarification.