* tramp.el (tramp-detect-ssh-controlmaster): New defun. (tramp-default-method): Use it.
* tramp.el (tramp-error-with-buffer): Don't show the connection
buffer when we are in completion mode.
(with-connection-property, tramp-completion-mode-p)
(tramp-action-process-alive, tramp-action-out-of-band)
(tramp-check-for-regexp, tramp-file-name-p, tramp-equal-remote)
(tramp-exists-file-name-handler): Fix docstring.
(with-progress-reporter): New defmacro.
(tramp-do-copy-or-rename-file, tramp-handle-dired-compress-file)
(tramp-maybe-open-connection): Use it.
(tramp-handle-copy-file): Add new optional parameter
`preserve-selinux-context'.
(tramp-file-name-for-operation): Add `set-file-selinux-context'.
(tramp-connectable-p): New defun.
(tramp-handle-expand-file-name)
(tramp-completion-handle-file-name-all-completions)
(tramp-completion-handle-file-name-completion): Use it.
(tramp-file-name-handler): Catch the error for some operations
when we are in completion mode.  This gives the user the chance to
correct the file name in the minibuffer.
(tramp-completion-mode-p): Use `non-essential'.

* tramp.el (tramp-file-name-for-operation):
Add file-selinux-context.