Correct some viewcvs links in docs and comments.
* doc/misc/cc-mode.texi: Remove reference to defunct viewcvs URL.
Instead, refer to the generic Savannah page, since it seems likely
that Savannah will never provide the equivalent service for GNU bazaar.
The other option would be to use the Git mirror URL, since it seems to
be being updated now, and of course browsing Git repos works just fine.
That would be a pretty poor advert for the GNU system though.

* etc/MH-E-NEWS: Update URL (viewcvs -> viewvc).

* lisp/net/tramp-fish.el: Update URL in comment (viewcvs -> viewvc).